requirements.txt
# Core Discord bot dependencies
discord.py==2.5.2
aiohttp==3.10.11

# Google API dependencies
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Utility dependencies
python-dotenv==1.0.1
python-dateutil==2.9.0.post0

# Async support
asyncio
